Add files via upload
This is the the Front end of the same task adder project. This is Task 3. There are 3 button. Add Task, Fetch Task and Delete Task. On clicking on any of the buttons the web page will navigate to the corresponding page and ask for the needed info. Once the info is given by the user, the required function lid adding, fetching or deleting will be implemented